refactor(breadcrumbs): add src and types folders
