+++ packages/breadcrumbs/index.d.ts
@@ -1,4 +1,2 @@
-import * as React from 'react';
-
-export { default } from './Breadcrumbs';
-export * from './Breadcrumbs';
+export { default } from './types/Breadcrumbs';
+export * from './types/Breadcrumbs';
