+++ packages/breadcrumbs/index.js
@@ -1,1 +1,1 @@
-export { default } from './Breadcrumbs';
+export { default } from './src/Breadcrumbs';
